requirements.txt
streamlit>=1.37.0
google-cloud-storage>=2.18.0
google-cloud-bigquery>=3.25.0
google-auth>=2.34.0
requests>=2.32.0
urllib3>=2.2.0
